Update baseball_data-checkpoint.ipynb
For some reason I always seem to get two separate changes and I can't commit them together, so this is the same thing: This is the updated code with markdown cells for all of the data cleaning we have done so far. Additionally I made a couple of changes to what we named a couple of columns and eliminated a couple of statistics (mostly just xFIP).